+++ api_v2/users_resource.py
@@ -1,5 +1,7 @@
+import datetime
+
 from flask import jsonify
-from flask_restful import Resource
+from flask_restful import Resource, abort
 from .parser_user import parser
 from data import db_session
 from data.users import User
@@ -9,12 +11,16 @@ class UsersResource(Resource):
     def get(self, users_id):
         session = db_session.create_session()
         users = session.get(User, users_id)
+        if not users:
+            abort(404)
         return jsonify({'users': users.to_dict(
-            only=('id', 'surname', 'name', 'age', 'position', 'speciality', 'address', 'email', 'hashed_password', 'modified_date'))})
+            only=('id', 'surname', 'name', 'age', 'position', 'speciality', 'address', 'email', 'modified_date', 'city_from'))})
 
     def delete(self, users_id):
         session = db_session.create_session()
-        users = session.query(User).get(users_id)
+        users = session.get(User, users_id)
+        if not users:
+            abort(404)
         session.delete(users)
         session.commit()
         return jsonify({'success': 'OK'})
@@ -24,7 +30,7 @@ class UsersListResource(Resource):
         session = db_session.create_session()
         users = session.query(User).all()
         return jsonify({'users': [item.to_dict(
-            only=('id', 'surname', 'name', 'age', 'position', 'speciality', 'address', 'email', 'hashed_password', 'modified_date')) for item in users]})
+            only=('id', 'surname', 'name', 'age', 'position', 'speciality', 'address', 'email', 'modified_date', 'city_from')) for item in users]})
 
     def post(self):
         args = parser.parse_args()
@@ -37,8 +43,8 @@ class UsersListResource(Resource):
             speciality=args['speciality'],
             address=args['address'],
             email=args['email'],
-            hashed_password=args['hashed_password'],
-            modified_date=args['modified_date']
+            modified_date=datetime.datetime.strptime(args['modified_date'], '%Y-%m-%d %H:%M:%S'),
+            city_from=args['city_from']
         )
         session.add(users)
         session.commit()
